chore: moved struct related code to its own file
fix: fixed error where it couldnt read &self, nor &mut self on methods
